lerobot/pyproject.toml
[project]
name = "lerobot"
version = "0.1.0"
description = "LeRobot in Lance format"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyarrow", "datasets", "pylance"]

[dependency-groups]
dev = ["ruff"]
